Add files via upload
The Three csv files were extracted in MYSQL database and the queries were performed on the basis of these data source. The Requests file contains the business requests and problem.
The answers.sql file contains the SQL queries which are answers to problem statements. The result png files contains the screenshot of the result.